Fix double closing carets when writing JSX
The jtsx mode command to close a JTSX element is a bit overzealous
about inserting a new closing caret `>`. Added a small bit of advice
that deletes the existing end caret if point is right before it. This
will probably break under some circumstances but it's a heck of a lot
better than having to delete that extra ending caret that is not at
point :p